test(support): move helper modules out of pytest's collection namespace
tests/test_orchestration.py and tests/test_data_utils.py define no tests. They
are support code imported by conftest.py and the slow tiers, but their test_*.py
names sit in pytest's collection path, so a genuine collection failure in them
is indistinguishable from normal deselection.

Move both into a tests/support/ package and update the three import sites plus
the documentation that names the old paths. Behaviour is unchanged: the same
functions with the same signatures, imported under a new path.

Refs #179
